# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meridian Advisory API",
    "description": "Consultation booking, payments, client and consultant portals, commerce and content APIs.\n\nAll responses use a single envelope: `{ success: true, data }` or `{ success: false, error: { code, message } }`.\n\nAuthenticate with `Authorization: Bearer <accessToken>`. Access tokens are short-lived; refresh them at `POST /api/auth/refresh`, which reads an httpOnly cookie.",
    "version": "1.0.0",
    "contact": {
      "name": "Support",
      "email": "support@meridianadvisory.co.ke"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/articles": {
      "get": {
        "summary": "List articles",
        "tags": [
          "Content"
        ],
        "description": "Anonymous callers see published articles only. A caller holding content.read may filter by status to see drafts and scheduled pieces.",
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "summary": "Create an article",
        "tags": [
          "Content"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/articles/{slug}": {
      "get": {
        "summary": "Article detail with related reading",
        "tags": [
          "Content"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/articles/{id}": {
      "patch": {
        "summary": "Update an article",
        "tags": [
          "Content"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "summary": "Archive an article",
        "tags": [
          "Content"
        ],
        "description": "Archives rather than deletes, so a published URL never 404s unexpectedly.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "development"
    }
  ],
...
}